remove wrapper for course form
